+++ Store/Users/UsersFetchDataAction.cs
@@ -8,7 +8,7 @@
 
         
 
-        public UsersFetchDataAction(string token, int searchPageNr, int itemsPerPage)
+        public UsersFetchDataAction(string token, int searchPageNr, long itemsPerPage)
         {
             Token = token;
             SearchPageNr = searchPageNr;
